+++ src/app/[local]/globals.css
