+++ src/app/[local]/layout.tsx
@@ -1,3 +1,16 @@
+import type { Metadata } from "next";
+import { Inter } from "next/font/google";
+import "./globals.css";
+import TrpcProvider from "@/lib/context/trpc";
+import { NextIntlClientProvider, useMessages } from "next-intl";
+
+const inter = Inter({ subsets: ["latin"] });
+
+export const metadata: Metadata = {
+  title: "Create Next App",
+  description: "Generated by create next app",
+};
+
 export default function LocaleLayout({
   children,
   params: { locale },
@@ -5,9 +18,16 @@ export default function LocaleLayout({
   children: React.ReactNode;
   params: { locale: string };
 }) {
+  const messages = useMessages();
   return (
     <html lang={locale}>
-      <body>{children}</body>
+      <body className={inter.className}>
+        <TrpcProvider>
+          <NextIntlClientProvider locale={locale} messages={messages}>
+            {children}
+          </NextIntlClientProvider>
+        </TrpcProvider>
+      </body>
     </html>
   );
 }
